+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [1.1.0] - 2023-10-25
-- Added signature verification method
+- Added symmetric HS256 signature verification method to `Jwt`
 - Added public key decoding to `KeyConverter`
 - Bumped `jjwt` version to 0.12.3
 - Bumped Kotlin version to 1.9.10